restructure algorithm
