--- solvers/qiskit/knapsack_quantum_tree_generator/knapsack_quantum_tree_generator.py
@@ -1,44 +1,0 @@
-from input import parse_input
-from run import run
-from run_solve import run_solve
-
-
-def _create_output(solve_results):
-    probabilities = solve_results["probabilities"]
-    knapsack = solve_results["knapsack"]
-
-    # Best feasible
-    best_feasible_bitstring = None
-    best_value = -1
-
-    for bitstring, _ in probabilities.items():
-        weight = sum(
-            knapsack.items[i].weight for i, b in enumerate(bitstring) if
-            b == '1')
-        if weight <= knapsack.capacity:
-            value = sum(
-                knapsack.items[i].value for i, b in enumerate(bitstring) if
-                b == '1')
-            if best_value == -1 or value > best_value:
-                best_value = value
-                best_feasible_bitstring = bitstring
-
-    if best_feasible_bitstring:
-        included_indexes = []
-        for i, bit in enumerate(best_feasible_bitstring):
-            if bit == '1':
-                included_indexes.append(knapsack.items[i].id)
-
-        # Sort indexes to match expected output if necessary,
-        # but the original code just appended them.
-        return f"{best_value}\n{included_indexes}"
-    else:
-        return "0\n[]"
-
-
-def _solve(input_data):
-    return _create_output(run_solve(parse_input(input_data)))
-
-
-if __name__ == "__main__":
-    run(_solve)
